Extract MetadataBuilder and UploadFiles from Process
This reduces the number of responsibilities that Process has
