+++ lib/sdr_client/deposit.rb
@@ -39,4 +39,6 @@ require 'sdr_client/deposit/files/direct_upload_response'
 require 'sdr_client/deposit/file'
 require 'sdr_client/deposit/file_set'
 require 'sdr_client/deposit/request'
+require 'sdr_client/deposit/upload_files'
+require 'sdr_client/deposit/metadata_builder'
 require 'sdr_client/deposit/process'
