+++ lib/sdr_client/deposit/process.rb
@@ -6,7 +6,6 @@ module SdrClient
   module Deposit
     # The process for doing a deposit
     class Process
-      BLOB_PATH = '/v1/direct_uploads'
       DRO_PATH = '/v1/resources'
       # @param [Request] metadata information about the object
       # @param [Class] grouping_strategy class whose run method groups an array of uploads
@@ -31,11 +30,12 @@ module SdrClient
 
       def run
         check_files_exist
-        file_metadata = collect_file_metadata
-        upload_responses = upload_file_metadata(file_metadata)
-        upload_files(upload_responses)
-        file_sets = build_filesets(uploads: upload_responses.values, files_metadata: files_metadata)
-        request = metadata.with_file_sets(file_sets)
+        upload_responses = UploadFiles.new(files: files, logger: logger, connection: connection).run
+        metadata_builder = MetadataBuilder.new(metadata: metadata,
+                                               grouping_strategy: grouping_strategy,
+                                               files_metadata: files_metadata,
+                                               logger: logger)
+        request = metadata_builder.with_uploads(upload_responses)
         upload_metadata(request.as_json)
       end
 
@@ -48,51 +48,6 @@ module SdrClient
         files.each do |file_name|
           raise Errno::ENOENT, file_name unless ::File.exist?(file_name)
         end
-      end
-
-      def collect_file_metadata
-        files.each_with_object({}) do |filename, obj|
-          obj[filename] = Files::DirectUploadRequest.from_file(filename)
-        end
-      end
-
-      # @param [Hash<String,Files::DirectUploadRequest>] file_metadata the filenames and their upload request
-      def upload_file_metadata(file_metadata)
-        Hash[file_metadata.map { |filename, metadata| [filename, direct_upload(metadata.to_json)] }]
-      end
-
-      def direct_upload(metadata_json)
-        logger.info("Starting an upload request: #{metadata_json}")
-        response = connection.post(BLOB_PATH, metadata_json, 'Content-Type' => 'application/json')
-        raise "unexpected response: #{response.inspect}" unless response.status == 200
-
-        logger.info("Response from server: #{response.body}")
-
-        Files::DirectUploadResponse.new(JSON.parse(response.body))
-      end
-
-      # @param [Hash<String,Files::DirectUploadResponse>] upload_responses the filenames and their upload response
-      def upload_files(upload_responses)
-        upload_responses.each do |filename, response|
-          upload_file(filename: filename,
-                      url: response.direct_upload.fetch('url'),
-                      content_type: response.content_type,
-                      content_length: response.byte_size)
-
-          logger.info('Upload complete')
-        end
-      end
-
-      def upload_file(filename:, url:, content_type:, content_length:)
-        logger.info("Uploading `#{filename}' to #{url}")
-
-        upload_response = connection.put(url) do |req|
-          req.body = ::File.open(filename)
-          req.headers['Content-Type'] = content_type
-          req.headers['Content-Length'] = content_length.to_s
-        end
-
-        raise "unexpected response: #{upload_response.inspect}" unless upload_response.status == 204
       end
 
       # @return [Hash<Symbol,String>] the result of the metadata call
